Update transactions and connections hash
Key type for both hashes was set to LibC::ULong which is set to UInt64 on Linux for 64-bit
systems. On Windows, LibC::ULong is set to UInt32. This causes an issue since object_id is
UInt64 value for 64-bit systems. This change fixes the issue by making sure that the key
type is UInt64 instead of relying on LibC::ULong.